move helper functions into spec folder and require all from rails_helper
