+++ spec/features/account_spec.rb
@@ -23,13 +23,4 @@ describe "user page" do
     expect(page.status_code).to be 404
   end
 
-def create_user_and_sign_in
-  user
-  visit '/'
-  click_link 'Sign in'
-  fill_in('user_email', with: 'test@localhost.co.uk')
-  fill_in('Password', with: 'password')
-  click_button 'Sign in'
-end
-
 end
