+++ spec/features/edit_dataset_spec.rb
@@ -1,5 +1,4 @@
 require "rails_helper"
-require_relative "datasets_spec_helper"
 
 describe 'editing datasets' do
   set_up
