+++ spec/support/log_in_helper.rb
@@ -1,0 +1,8 @@
+def create_user_and_sign_in
+  user
+  visit '/'
+  click_link 'Sign in'
+  fill_in('user_email', with: 'test@localhost.co.uk')
+  fill_in('Password', with: 'password')
+  click_button 'Sign in'
+end
